requirements.txt
asgiref==3.5.2
bitly-api-py3==0.3
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
djangorestframework==3.13.1
idna==3.3
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
stripe==3.5.0
tzdata==2022.1
urllib3==1.26.10
